feat: Added DataBasePresenter and TreePresenter implementation
